replace SynchronizedViewList backed uses AlternateIndexList
